NTP-838: Refactored the code and included all javascript files through the single webpack file - Ensure JS is being conditionally applied to page element using data-module - Use Axios instead of Fetch for more backwards compatibility
