+++ UI/package.json
@@ -24,6 +24,7 @@
   "homepage": "https://github.com/DFE-Digital/find-a-tuition-partner#readme",
   "dependencies": {
     "@ministryofjustice/frontend": "^1.4.2",
+    "axios": "^1.2.5",
     "govuk-frontend": "^4.1.0"
   },
   "devDependencies": {
